+++ src/components/ui/Heading.tsx
@@ -1,5 +1,5 @@
 import { CSSProperties, PropsWithChildren } from 'react';
-import { SpacingProperties, spacingsToClasses } from './helpers';
+import { SpacingProperties, spacingsToClasses } from './helpers/spacings';
 import { LegacyColor, legacyColors } from './helpers/colors';
 
 type HeadingType = 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6';
